backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-pptx==0.6.23
openpyxl==3.1.2
lxml==5.2.2
pymysql==1.1.1
pydantic==2.7.1
apscheduler==3.10.4
